feat: improve error message when trying to use disabled plugin feature
The plugin feature can be disabled at compile-time. Previously,
attempting to use the `--client-plugins-dir` or `--mgmt-plugins-dir`
flags without plugin support compiled into the wfx binary resulted
in a generic "unknown flags" error. This commit introduces a more
informative error message, explicitly stating that wfx was built
without plugin support, improving clarity for end users and aiding
in troubleshooting.
